cnslk

+++ ft_print.h
@@ -12,7 +12,7 @@ int			ft_printf(const char *fmt_or_nch, ...);
 size_t		ft_putchar(char inputchar_num);
 size_t		ft_putchar_string(char *sub_string);
 size_t   	ft_putaddr_to_hexia(unsigned long long num, char *basestring);
-size_t	    ft_putnum_to_hexia(unsigned int num, char *basestring);
+size_t   	ft_putnum_to_hexia(long long num, char *basestring);
 size_t		inputnbr(long long subject, char* src);
 
 #endif
